requirements.txt
Flask==3.0.0
flask-cors==4.0.0
Werkzeug==3.0.1
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
PyJWT==2.8.0
reportlab==4.0.7
